Setup Todo MVC in React + Rails 5.1
Building off previous work to correctly install React within the Rails
5.1 environment, this PR represents the very first steps required to
build components the Launchpad Lab way.

The setup generally occurs in four parts:

 1. Wrap our component rendering in route definitions
  - For now, our Route definitions occur at the top level of MainApp.
 2. Define our 'todo' structure
  - For every section of our MainApp, we want a folder like 'todo'.
  - Todo might look like this:

    main
    .
    | |____todo
    | | |____actions.js
    | | |____index.js
    | | |____reducer.js
    | | |____components
    | | | |____footer.js
    | | | |____header.js
    | | | |____index.js
    | | | |____item.js
    | | | |____list.js
    | | |____views
    | | | |____index.js
    | | | |____todo-show.js

  3. For each imagined component, we want to define each component and
    establish import/exports.
  4. Allow the components to render dummy data, to ensure React is
    rendering.

The last bit of work was initializing a store with `redux`, using
reducers to define the shape of the store object, then creating actions.

Ultimately this all leads to mapStateToProps and mapDispatchToProps -
allowing us to change the state of our components with event handlers.